requirements.txt
MarkupSafe==3.0.2 #
PyJWT==2.10.0 #
PyMuPDF==1.23.6 #
PyMuPDFb==1.23.6 #
PyYAML==6.0.2 #
SQLAlchemy==2.0.36 #
Werkzeug==3.1.3 #
XlsxWriter==3.2.0 #
aiohappyeyeballs==2.4.3 #
aiohttp==3.11.7 #
aiosignal==1.3.1 #
altair==4.0.0 #
annotated-types==0.7.0 #
anyio==4.6.2.post1 #
attrs==24.2.0 #
azure-ai-documentintelligence==1.0.0b3 #
azure-ai-formrecognizer==3.3.1 #
azure-ai-translation-text==1.0.1 #
azure-common==1.1.28 #
azure-core==1.32.0 #
azure-identity==1.14.1 #
azure-search==1.0.0b2 #
azure-search-documents==11.4.0b11 #
azure-storage-blob==12.20.0 #
beautifulsoup4==4.12.3 #
blinker==1.9.0 #
certifi==2024.8.30 #
cffi==1.17.1 #
charset-normalizer==3.4.0 #
click==8.1.7 #
cryptography==43.0.3 #
dataclasses-json==0.6.7 #
distro==1.9.0 #
docstring-parser==0.16 #
docx2txt==0.8 #
docx2txt2==1.0.4 #
entrypoints==0.4 #
et-xmlfile==2.0.0 #
faiss-cpu==1.7.4 #
flask==3.1.0 #
flask-cors==5.0.0 #
frozenlist==1.5.0 #
greenlet==3.1.1 #
gunicorn==23.0.0 #
h11==0.14.0 #
httpcore==1.0.7 #
httpx==0.27.2 #
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.43
langchain-text-splitters==0.2.4
langchain_openai==0.1.20
langdetect==1.0.9
langsmith==0.1.146
openai==1.35.7
openpyxl==3.1.5
pandas==2.2.3
pillow==10.1.0
pydantic==2.6.1
pydantic-core==2.16.2
pypdf==4.2.0
pypdf2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
rank_bm25==0.2.2
regex==2023.10.3
requests==2.31.0
threadpoolctl==3.5.0
tiktoken==0.8.0
typing-extensions==4.12.2
typing-inspect==0.9.0
urllib3==2.2.3
uvicorn==0.32.1
waitress==3.0.2
xlrd==2.0.1
yarl==1.18.0